requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
channels==4.2.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.2.1
django-allauth==65.9.0
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-haystack==3.3.0
django-health-check==3.18.3
django-otp==1.6.0
django-silk==5.4.0
django-storages==1.14.6
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
drf-yasg==1.21.10
gprof2dot==2025.4.14
inflection==0.5.1
kombu==5.5.4
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
